Replace images with ImagesApi class
