+++ agent/package.json
@@ -63,7 +63,8 @@
       "no-plusplus": "off",
       "no-restricted-syntax": "off",
       "@typescript-eslint/ban-ts-comment": "off",
-      "import/prefer-default-export": "off"
+      "import/prefer-default-export": "off",
+      "class-methods-use-this": "off"
     }
   },
   "prettier": {
